changed ledgacy - long instr_ptr type to size_t type, this was causing an issue on win arm64

+++ core/vm/arch/jit/arm64/jit_arm_a64.h
@@ -788,7 +788,7 @@ namespace Runtime {
    * Prototype for jit function
    */
   typedef long (*jit_fun_ptr)(long cls_id, long mthd_id, size_t* cls_mem, size_t *inst,
-                              size_t *op_stack, long *stack_pos, StackFrame **call_stack,
+                              size_t *op_stack, size_t*stack_pos, StackFrame **call_stack,
                               long *call_stack_pos, size_t **jit_mem, long *offset, int64_t*ints);
   
   
@@ -802,7 +802,7 @@ namespace Runtime {
     static void Initialize(StackProgram* p);
     
     // Executes machine code
-    long Execute(StackMethod* method, size_t* inst, size_t* op_stack, long* stack_pos,
+    long Execute(StackMethod* method, size_t* inst, size_t* op_stack, size_t* stack_pos,
                  StackFrame** call_stack, long* call_stack_pos, StackFrame* frame);
   };
 }
